+++ cli/integration-test/sample-test/src/sample_test/declaration/a.clj
@@ -1,0 +1,3 @@
+(ns sample-test.declaration.a)
+
+(def some-var 1)